+++ CHANGELOG.md
@@ -246,7 +246,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
   - BerRef::eoc()
   - Der::from\_vec\_unchecked()
 - Create struct `IdNumber`, and change the following functions and methods type.
-  - The return type of `IdRef::number()`
+  - The return type of `IdRef::number`
+  - The argument type of `Id::new`
 
 ### Changed